Split VirtualList tests into three files
Enact-DCO-1.0-Signed-off-by: Roy Sutton roy.sutton@lge.com
